+++ tests/unit/EntityManagerFactoryTest.php
@@ -3,6 +3,7 @@ declare(strict_types = 1);
 
 namespace Tests\Unit;
 
+use Doctrine\Common\Annotations\AnnotationException;
 use PHPUnit\Framework\TestCase;
 use SignNow\Rest\EntityManager;
 use SignNow\Rest\EntityManagerFactory;
@@ -17,12 +18,14 @@ use SignNow\Rest\EntityManagerFactory;
  * @uses \SignNow\Rest\EntityManager
  * @uses \SignNow\Rest\EntityManager\AnnotationResolver
  * @uses \SignNow\Rest\Service\Request\Pool
+ * @uses \SignNow\Rest\Service\Serializer\Handler\SplFileInfo
  * @uses \SignNow\Rest\Service\Serializer\Handler\File
+ * @uses \SignNow\Rest\Service\Serializer\Handler\FileLink
  */
 class EntityManagerFactoryTest extends TestCase
 {
     /**
-     * @throws \Doctrine\Common\Annotations\AnnotationException
+     * @throws AnnotationException
      *
      * @covers ::createEntityManager
      */