+++ CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2021-06-18
+
+### Added
+
+- Add new file entity to serialize
+
+### Changed
+
+- Refactor factories
+- Update serializer version
+- Update guzzle to 7+ version
+
+### Deprecated
+
+- Deprecate SignNow\Rest\EntityManagerFactory
+
+### Fixed
+
+- Fix array type mapping
+- Fix FileLink type formatting for multipart
+- Fix SplFileInfo type formatting for multipart
+
+## [v1.0.1] - 2021-04-16
+
+### Fixed
+
+- Add required php extension to composer.json
+
 ## [v1.0.0] - 2019-10-23
 
 ### Added